deps/v8/test/inspector/debugger/breakpoints-expected.txt

Checks breakpoints.

Running test: testRemoveBreakpoint
Debugger.removeBreakpoint when agent is disabled:
{
    error : {
        code : -32000
        message : Debugger agent is not enabled
    }
    id : <messageId>
}
Remove breakpoint with invalid breakpoint id:
{
    id : <messageId>
    result : {
    }
}
{
    id : <messageId>
    result : {
    }
}

Running test: testSetBreakpointByUrl
Adding conditional (arg === 1) breakpoint
evaluating foo1(0):
  not paused
evaluating foo1(1):
  hit expected breakpoint

Evaluating another script with the same url
evaluating foo2(0):
  not paused
evaluating foo2(1):
  hit expected breakpoint

Removing breakpoint
evaluating foo1(1):
  not paused
evaluating foo2(1):
  not paused

Adding breakpoint back
evaluating foo1(0):
  not paused
evaluating foo1(1):
  hit expected breakpoint

Disabling debugger agent
evaluating foo1(1):
  not paused
evaluating foo2(1):
  not paused

Enabling debugger agent
evaluating foo1(1):
  not paused
evaluating foo2(1):
  not paused

Running test: testSetBreakpointInScriptsWithDifferentOffsets
Adding breakpoint
evaluating foo1(0):
  hit expected breakpoint
evaluating foo2(0):
  hit expected breakpoint
